postureservice small bug fix
fixed minor bug by switching interval value to double instead of long to house decimal places

+++ app/src/main/java/com/example/csci3310project/PostureService.java
@@ -63,7 +63,7 @@ public class PostureService extends LifecycleService {
     private ScheduledExecutorService scheduler;
     private int consecutiveBadPosture = 0;
     private double threshold;
-    private long interval;
+    private double interval;
     private ProcessCameraProvider cameraProvider;
     private ImageCapture imageCapture;
     private SensorManager sensorManager;
@@ -294,7 +294,7 @@ public class PostureService extends LifecycleService {
         return pitch;
     }
 
-    private long parseInterval(String intervalStr) {
+    private double parseInterval(String intervalStr) {
         switch (intervalStr) {
             case "Real-time": return 500;
             case "5 seconds": return 5000;
